Moved conversion methods to a utils class
