+++ src/main/java/ch/heigvd/sitr/gui/simulation/Displayer.java
@@ -8,7 +8,7 @@ package ch.heigvd.sitr.gui.simulation;
 import java.awt.Graphics2D;
 
 /**
- * Interface Displayer implements some methode for GUI representation
+ * Displayer interface describes methods for GUI representation
  *
  * @author Luc Wachter, Loris Gilliand
  */
