+++ src/main/java/ch/heigvd/sitr/model/Simulation.java
@@ -119,70 +119,6 @@ public class Simulation {
     }
 
     /**
-     * Convert meters per second to kilometers per hour
-     *
-     * @param mps The amount of m/s to convert
-     * @return the corresponding amount of km/h
-     */
-    public static double mpsToKph(double mps) {
-        // m/s => km/h : x * 3.6
-        return mps * 3.6;
-    }
-
-    /**
-     * Convert kilometers per hour to meters per second
-     *
-     * @param kph The amount of km/h to convert
-     * @return the corresponding
-     */
-    public static double kphToMps(double kph) {
-        // km/h => m/s : x / 3.6
-        return kph / 3.6;
-    }
-
-    /**
-     * Convert m to px
-     *
-     * @param scale the ratio px/m
-     * @param m     the number of m
-     * @return the number of px
-     */
-    public static int mToPx(double scale, double m) {
-        return (int) Math.round(m * scale);
-    }
-
-    /**
-     * Convert px to m
-     *
-     * @param scale the ratio px/m
-     * @param px    the number of px
-     * @return the number of px
-     */
-    public static double pxToM(double scale, int px) {
-        return px / scale;
-    }
-
-    /**
-     * Convert m to px
-     *
-     * @param m the number of m
-     * @return the number of px
-     */
-    public int mToPx(double m) {
-        return Simulation.mToPx(scenario.getScale(), m);
-    }
-
-    /**
-     * Convert px to m
-     *
-     * @param px the number of px
-     * @return the number of px
-     */
-    public double pxToM(int px) {
-        return Simulation.pxToM(scenario.getScale(), px);
-    }
-
-    /**
      * Parse the OpenDrive XML file
      *
      * @param roadNetwork       The Road network that will contains OpenDrive road network
