+++ src/main/java/ch/heigvd/sitr/vehicle/VehicleRenderer.java
@@ -6,6 +6,7 @@
 package ch.heigvd.sitr.vehicle;
 
 import ch.heigvd.sitr.model.Simulation;
+import ch.heigvd.sitr.utils.Conversions;
 
 import java.awt.*;
 
@@ -39,7 +40,8 @@ public class VehicleRenderer {
         // TODO store drawing information in Vehicle (x, y, length, width, color)
 //        g.setColor(vehicle.getVehicleController().getColor());
         g.setColor(Color.BLUE);
-        g.fillRect((int) Simulation.mToPx(scale, vehicle.getPosition()), 20, Simulation.mToPx(scale, vehicle.getLength()), Simulation.mToPx(scale, vehicle.getWidth()));
+        g.fillRect(Conversions.mToPx(scale, vehicle.getPosition()), 20, Conversions.mToPx(scale,
+                vehicle.getLength()), Conversions.mToPx(scale, vehicle.getWidth()));
     }
 
     /**
