+++ src/test/java/ch/heigvd/sitr/model/SimulationTest.java
@@ -25,16 +25,4 @@ class SimulationTest {
         HashMap<VehicleControllerType, Integer> map = new HashMap<>();
         simulation = new Simulation(ScenarioType.SIMPLE_ROAD, VehicleBehaviourType.STOP, map);
     }
-
-    @Test
-    public void shouldReturnCorrectAmountOfKph() {
-        assertEquals(3.6, Simulation.mpsToKph(1));
-        assertEquals(72.0, Simulation.mpsToKph(20));
-    }
-
-    @Test
-    public void shouldReturnCorrectAmountOfMps() {
-        assertEquals(1.0, Simulation.kphToMps(3.6));
-        assertEquals(20.0, Simulation.kphToMps(72));
-    }
 }
